t2

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -1,4 +1,3 @@
-#include "main.h"
 #include <stdio.h>
 
 /**
@@ -6,7 +5,7 @@
  * @str: string to get length
  * Return: string length
  */
-int _strlen(const char *str)
+size_t _strlen(const char *str)
 {
 	size_t l = 0;
